Day9 1st pass, cheated a bit on extending the tape

+++ src/day5.rs
@@ -9,13 +9,13 @@ pub fn part1()
     file.read_to_string(&mut file_data).expect("Unable to read file");
 
     let mut program = intcode::Program::from_tape(file_data.split(",")
-                                                .map(|num| num.parse::<i32>())
+                                                .map(|num| num.parse::<i64>())
                                                 .filter(|num| num.is_ok())
                                                 .map(|num| num.unwrap())
                                                 .collect());
     program.push_input(1);
 
-    println!("Day 5 Part 1: {}", program.filter(|n| *n > 0).collect::<Vec<i32>>()[0]);
+    println!("Day 5 Part 1: {}", program.filter(|n| *n > 0).collect::<Vec<i64>>()[0]);
 
 }
 
@@ -26,11 +26,11 @@ pub fn part2()
     file.read_to_string(&mut file_data).expect("Unable to read file");
 
     let mut program = intcode::Program::from_tape(file_data.split(",")
-                                                .map(|num| num.parse::<i32>())
+                                                .map(|num| num.parse::<i64>())
                                                 .filter(|num| num.is_ok())
                                                 .map(|num| num.unwrap())
                                                 .collect());
     program.push_input(5);
 
-    println!("Day 5 Part 2: {}", program.filter(|n| *n > 0).collect::<Vec<i32>>()[0]);
+    println!("Day 5 Part 2: {}", program.filter(|n| *n > 0).collect::<Vec<i64>>()[0]);
 }
