+++ src/day7.rs
@@ -9,8 +9,8 @@ pub fn part1()
     let mut file_data = String::new();
     file.read_to_string(&mut file_data).expect("Unable to read file");
 
-    let prog_data : Vec<i32> = file_data.split(",")
-                                .map(|num| num.parse::<i32>())
+    let prog_data : Vec<i64> = file_data.split(",")
+                                .map(|num| num.parse::<i64>())
                                 .filter(|num| num.is_ok())
                                 .map(|num| num.unwrap())
                                 .collect();
@@ -43,8 +43,8 @@ pub fn part2()
     let mut file_data = String::new();
     file.read_to_string(&mut file_data).expect("Unable to read file");
 
-    let prog_data : Vec<i32> = file_data.split(",")
-                                .map(|num| num.parse::<i32>())
+    let prog_data : Vec<i64> = file_data.split(",")
+                                .map(|num| num.parse::<i64>())
                                 .filter(|num| num.is_ok())
                                 .map(|num| num.unwrap())
                                 .collect();
